Fix python3-native dependency in bl808-firmware
This change removes the RDEPENDS from bl808-firmware because there are
no runtime dependencies for the target from this task. This ensures that
python is not compiled for the target when not needed.

It also removes python3-native from the DEPENDS list because the
python3native.bbclass already includes that dependency.

The line at the top of mergebin.py was modified to call the python3 in
PATH in case that differs from the system installed python3 for some
reason.