Fix do_install dependencies for bl808-firmware
Move the dependencies on opensbi and u-boot to do_install[depends]
because the default DEPENDS line does not depend on do_deploy from
those tasks. DEPENDS sets a dependency on do_populate_sysroot which may
be completed before do_deploy. This leaves a possible race condition
where binaries used in do_install are not populated yet. This change
specifies that do_install depends on do_deploy from those dependencies
to resolve the race.